Fix integer overflows in BIBOP GC on I32LP64 platforms.
On my AMD64 box, (make-vector (expt 2 N)) was crashing the process with
SIGSEGV, for N from 29 to 53 inclusive.  Now it crashes the process with
SIGABRT, because the GC aborts rather than returning NULL (as the VM expects it
to) when it cannot allocate more memory.  Yay.

+++ c/bibop/areas.h
@@ -68,7 +68,7 @@ typedef struct Space {
 
 
 /* Allocate an area of between MINIMUM and MAXIMUM pages, inclusive. */
-extern Area* s48_allocate_area(unsigned int minimum, unsigned int maximum,
+extern Area* s48_allocate_area(unsigned long minimum, unsigned long maximum,
 			       unsigned char generation_index, area_type_size_t area_type_size);
 
 /* Remove AREA from the list starting with START */
